requirements.txt
Flask
Pillow
tensorflow
numpy
pandas
joblib
sklearn
matplotlib
tqdm
h5py
